AB#feature/8776-mapping fix naming
